[issue #660][feature]: remove transactional annotation from interfaces Move more classes from internal-api to util
