+++ agent-commons/pom.xml
@@ -87,11 +87,26 @@
                 </exclusion>
                 <exclusion>
                     <groupId>org.squeryl</groupId>
-                    <artifactId>squeryl_2.9.0</artifactId>
+                    <artifactId>squeryl_2.10</artifactId>
                 </exclusion>
                 <exclusion>
-                    <groupId>org.springframework</groupId>
-                    <artifactId>spring-webmvc</artifactId>
+                    <groupId>com.thoughtworks.xstream</groupId>
+                    <artifactId>xstream</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+
+        <dependency>
+            <groupId>com.griddynamics.genesis</groupId>
+            <artifactId>util</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.squeryl</groupId>
+                    <artifactId>squeryl_2.10</artifactId>
+                </exclusion>
+               <exclusion>
+                    <groupId>com.griddynamics.genesis</groupId>
+                    <artifactId>api</artifactId>
                 </exclusion>
                 <exclusion>
                     <groupId>org.springframework</groupId>
@@ -105,24 +120,12 @@
                     <groupId>javax.transaction</groupId>
                     <artifactId>jta</artifactId>
                 </exclusion>
-                <exclusion>
-                    <groupId>org.jclouds.api</groupId>
-                    <artifactId>nova</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>net.liftweb</groupId>
-                    <artifactId>${net.liftweb.lift-json.id}</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>com.jcraft</groupId>
-                    <artifactId>jsch</artifactId>
+                 <exclusion>
+                    <groupId>net.sf.ehcache</groupId>
+                    <artifactId>ehcache-core</artifactId>
                 </exclusion>
 
             </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>com.griddynamics.genesis</groupId>
-            <artifactId>util</artifactId>
         </dependency>
      </dependencies>
