+++ genesis-shell/pom.xml
@@ -62,6 +62,11 @@
             <groupId>net.liftweb</groupId>
             <artifactId>${net.liftweb.lift-json.id}</artifactId>
         </dependency>
+        <!-- For spring-context-->
+        <dependency>
+            <groupId>commons-logging</groupId>
+            <artifactId>commons-logging</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
