+++ internal-api/src/main/scala/com/griddynamics/genesis/common/CRUDService.scala
@@ -23,20 +23,13 @@
 
 package com.griddynamics.genesis.common
 
-import org.springframework.transaction.annotation.Transactional
 import com.griddynamics.genesis.api.{Failure, ExtendedResult}
 
-
 trait CRUDService[A, KEY] {
-    @Transactional(readOnly = true)
     def list: Seq[A]
-    @Transactional(readOnly = false)
     def create(a: A): ExtendedResult[A] = Failure(compoundServiceErrors = Seq("This service cannot create objects"))
-    @Transactional(readOnly = false)
     def update(a: A) : ExtendedResult[A] = Failure(compoundServiceErrors = Seq("This service cannot update objects"))
-    @Transactional(readOnly = false)
     def delete(a: A) : ExtendedResult[A] = Failure(compoundServiceErrors = Seq("This service cannot update objects"))
-    @Transactional(readOnly = true)
     def get(key: KEY) : Option[A]
 }
