+++ internal-api/src/main/scala/com/griddynamics/genesis/service/RemoteAgentsService.scala
@@ -25,9 +25,7 @@ package com.griddynamics.genesis.service
 
 import com.griddynamics.genesis.common.CRUDService
 import com.griddynamics.genesis.api.RemoteAgent
-import org.springframework.transaction.annotation.Transactional
 
 trait RemoteAgentsService extends CRUDService[RemoteAgent, Int]{
-    @Transactional(readOnly = true)
     def findByTags(tags: Seq[String]): Seq[RemoteAgent]
 }
