+++ pom.xml
@@ -584,7 +584,6 @@
                 <version>${org.springframework.version}</version>
             </dependency>
 
-
             <dependency>
                 <groupId>org.springframework</groupId>
                 <artifactId>spring-test</artifactId>
