+++ run-local/pom.xml
@@ -40,12 +40,47 @@
         <dependency>
             <groupId>com.griddynamics.genesis</groupId>
             <artifactId>internal-api</artifactId>
+           <exclusions>
+                <exclusion>
+                    <groupId>com.griddynamics.genesis</groupId>
+                    <artifactId>api</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>org.squeryl</groupId>
+                    <artifactId>squeryl_2.10</artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
+
 
         <dependency>
             <groupId>com.griddynamics.genesis</groupId>
             <artifactId>util</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>com.griddynamics.genesis</groupId>
+                    <artifactId>api</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>org.springframework</groupId>
+                    <artifactId>spring-jdbc</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>org.springframework</groupId>
+                    <artifactId>spring-tx</artifactId>
+                </exclusion>
+                <exclusion>
+                    <groupId>javax.transaction</groupId>
+                    <artifactId>jta</artifactId>
+                </exclusion>
+                 <exclusion>
+                    <groupId>net.sf.ehcache</groupId>
+                    <artifactId>ehcache-core</artifactId>
+                </exclusion>
+
+            </exclusions>
         </dependency>
+
 
         <dependency>
             <groupId>ch.qos.logback</groupId>
