+++ util/src/main/java/com/griddynamics/genesis/plugin/adapter/AbstractSimpleActionExecutor.java
