+++ util/src/main/java/com/griddynamics/genesis/plugin/adapter/AbstractSimpleSyncActionExecutor.java
