+++ util/src/main/java/com/griddynamics/genesis/plugin/utils/ScalaUtils.java
