+++ util/src/main/scala/com/griddynamics/genesis/util/StringUtils.scala
